Fix windows utf16 bug
I had originally stored windows passwords as a blob from a Utf8 string.

Since the windows vault api allows the application to set the encoding,
this worked fine when the keyring was the only application using and
editing the password

However, editing by a third party application (e.g. through the windows
vault interface) would introduce a Utf16 string as the password.
Decoding that password as utf8 would then result in a wrong output.

The solution is to convert all password strings to utf16 before storing
as a blob. Since I expect any other third party applications to use utf16,
this should be a global fix for the issue.

To that end:

- add byteorder as dependency on windows
- For encoding password, first encode as windows utf16, then convert
  to vec<u16>, then convert to vec<u8> using byteorder, then pass to
  the windows api.
- For decoding, first get u8 slice, then convert to Vec<u16>, then read
  to String using from_utf16.

Note that Windows u16 is LE