fix osx for new security-framework
when bringing in pull request for security-framework support,
unfortunately the dependency was not updated from the git branch, so
unfortunately that was committed.

Then, I updated the dependecy to the public 0.3.0, and didn't check
it, so commits after that don't even compile (fortunately, it's only
moving the cli from bin too examples)

So, this fixes the osx section.

All tests pass, make sure RUST_TEST_THREADS=1