Linux: Only include application name on create
Currently the get_password method is doing a strict search including the
application name that was sent in during the Struct initialiation.  This
means that attempting to query for any password that was not previously
created by this app specifically will fail.

We adjust this to only create the application field on creation, and
leave searching to just look for service and username.

This brings "better" parity to the python and rust clients, as well as
parity within the MacOS and Linux libraries.