Don't allow empty strings to name credentials.
Fixes hwchen/keyring-rs#86.

The fix, since we can't fail to create entries without modifying the API, is to make sure the created entries with empty strings can't be used to set or get passwords.  We do this by introducing an "always invalid" platform credential that we create for these entries.

Updates docs to clarify that the service and user names must be non-empty.