Improve docs, bump minor version number.
Based on review discussion, this change should really be considered a minor version change rather than a patch release.

The docs have been updated to note the change in behavior and the workaround available to anyone who was relying on the prior behavior.  In addition, a warning has been issued that the next major release may change the `Entry` creation API to be fallible.