Remove patch versions in Cargo.toml
Since we are using semver dependencies, and we keep the Cargo.lock file in the project, the patch versions are actually irrelevant and a bit misleading (because Cargo.lock is often newer).

This checkin also includes a cargo update, but there have been no changes since the last release so there's no need for a new release.