Add service-level search in secret service.
This adds an option to the keystore constructor to specify whether to use collection-level or service-level search.

This also reverts secret service to v2, so we can release, with TODOs that show what has to change for v3.

(Ironically, in secret-service v2, service-level search returns everything in on list, rather than separating locked from unlocked, so it's not at all clear that service-level search gets us anything.)

This compiles but there are no new tests and the tests have never been run.