Bump version to 3.1.0.
This adds API to credentials to provide better debugging, and implements that API in the built-in credential stores so it's exposed in the CLI.

This is a dot release rather than a full release because the additional API is implemented in a backward-compatible way. Clients *can* extend their implementations to take advantage of it, but they don't have to.